requirements/base.txt
Django>=3.0,<4.0
djangorestframework>=3.11
markdown>=3
django-cors-headers>=3.4
psycopg2-binary>=2.8
